Make sure the account assigned to a player still is configured - otherwise fall back to the first existing account.
This is an issue I've faced with TIDAL, too: a user has multiple accounts in LMS. Then removes one. But it remains assigned to a player. The old code would continue to try to use that now removed account, and there's no way to select the remaining account. With my change in this situation the player would automatically get the remaining account assigned.

Re-use the newly introduced `getUseradata()` in some other places where we need to access a user's account.